alatanapizza-master/src/com/internousdev/alatanapizza/dto/ChangePasswordDTO.java
package com.internousdev.alatanapizza.dto;

public class ChangePasswordDTO {
	private String userid;
	private String password;
	private String answer;
	private int question;
	public String getUserid() {
		return userid;
	}
	public void setUserid(String userid) {
		this.userid = userid;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}
	public String getAnswer() {
		return answer;
	}
	public void setAnswer(String answer) {
		this.answer = answer;
	}
	public int getQuestion() {
		return question;
	}
	public void setQuestion(int question) {
		this.question = question;
	}

	}



